+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+[L10n]
+* New command `l10n cleanup_xml` to remove unused strings from XML files.
+
 ## 1.6.22
 
 [L10n]